// src/bcp-client/src/app/api/model/inscriptionCommand.ts
/**
 * BilliardClubParisien.Api
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 0.2.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { InscriptionImageCommand } from './inscriptionImageCommand';


export interface InscriptionCommand { 
    id?: number;
    firstName: string;
    lastName: string;
    sex: string;
    email: string;
    phone?: string | null;
    isMemberBefore?: boolean;
    formula?: string | null;
    joinCompetition?: boolean;
    competitionCats?: Array<string> | null;
    motivation?: string | null;
    status: string;
    inscriptionImages?: Array<InscriptionImageCommand> | null;
}